+++ packages/sdks/output/react/CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.3.1
+
+- Feature: Added SDK version to data sent to visual editor for improved debugging.
+- Fix: Columns block: removed redundant margin-left in first column.
+- Fix: dynamic action bindings in repeated data.
+
+### 0.3.0
+
+- No Changes.
+
 ### 0.2.3
 
 - Feature: Added TS types to package.